Fix CLIP layers dtype for GPU compatibility
Revert dtype='int32' back to dtype=int to maintain GPU support.
The backend needs to determine the appropriate integer dtype for
correct device placement across different frameworks (TF, JAX, PyTorch).

+++ keras_hub/src/models/clip/clip_layers.py
@@ -52,7 +52,7 @@ class CLIPVisionEmbedding(layers.Layer):
         self.position_ids = self.add_weight(
             shape=(1, self.num_positions),
             initializer="zeros",
-            dtype="int32",
+            dtype=int,
             trainable=False,
             name="position_ids",
         )
